pyproject.toml
[tool.poetry]
name = "data-loader-elt"
version = "0.1.0"
description = ""
authors = ["angelina <cher.rly2036@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.10.4"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
indent-width = 4
fix = true


[tool.ruff.lint]
select = ["E", "W", "C", "F", "N", "D", "I"]
preview = true
ignore = ["D100", "D104", "D203", "D213"]
